[chore] Move HFSentenceTransformerRetriever to huggingface module for consistency
