+++ docs/api_reference/retrievers/huggingface.md
@@ -1,6 +1,6 @@
 <!-- markdownlint-disable-file MD041 -->
 
-::: src.fed_rag.retrievers.hf_sentence_transformer
+::: src.fed_rag.retrievers.huggingface.hf_sentence_transformer
     options:
       members:
         - HFSentenceTransformerRetriever
