+++ examples/ra-dit/ra_dit/retrievers/dragon.py
@@ -1,6 +1,6 @@
 """Dragon Retriever."""
 
-from fed_rag.retrievers.hf_sentence_transformer import (
+from fed_rag.retrievers.huggingface.hf_sentence_transformer import (
     HFSentenceTransformerRetriever,
 )
