+++ tests/retrievers/test_hf_sentence_transformer.py
@@ -7,7 +7,7 @@ from sentence_transformers import SentenceTransformer
 
 from fed_rag.base.retriever import BaseRetriever
 from fed_rag.exceptions import MissingExtraError
-from fed_rag.retrievers.hf_sentence_transformer import (
+from fed_rag.retrievers.huggingface.hf_sentence_transformer import (
     HFSentenceTransformerRetriever,
     InvalidLoadType,
     LoadKwargs,
@@ -15,9 +15,7 @@ from fed_rag.retrievers.hf_sentence_transformer import (
 
 
 def test_hf_sentence_transformer_retriever_class() -> None:
-    names_of_base_classes = [
-        b.__name__ for b in HFSentenceTransformerRetriever.__mro__
-    ]
+    names_of_base_classes = [b.__name__ for b in HFSentenceTransformerRetriever.__mro__]
     assert BaseRetriever.__name__ in names_of_base_classes
 
 
@@ -128,7 +126,7 @@ def test_hf_sentence_transformer_retriever_class_init_encoder(
     assert retriever.context_encoder is None
 
 
-@patch("fed_rag.retrievers.hf_sentence_transformer.SentenceTransformer")
+@patch("fed_rag.retrievers.huggingface.hf_sentence_transformer.SentenceTransformer")
 def test_load_model_from_hf_constructs_sentence_transformer_obj(
     mock_sentence_transformer: MagicMock,
 ) -> None:
@@ -139,14 +137,12 @@ def test_load_model_from_hf_constructs_sentence_transformer_obj(
     )
 
     # assert
-    mock_sentence_transformer.assert_called_once_with(
-        "fake_name", device="cpu"
-    )
+    mock_sentence_transformer.assert_called_once_with("fake_name", device="cpu")
     assert retriever.query_encoder is None
     assert retriever.context_encoder is None
 
 
-@patch("fed_rag.retrievers.hf_sentence_transformer.SentenceTransformer")
+@patch("fed_rag.retrievers.huggingface.hf_sentence_transformer.SentenceTransformer")
 def test_load_model_from_hf_constructs_sentence_transformer_obj_dual(
     mock_sentence_transformer: MagicMock,
 ) -> None:
